requirements.txt
certifi==2020.6.20
chardet==3.0.4
Django==3.1.4
django-ckeditor==6.0.0
django-crispy-forms==1.10.0
django-currentuser==0.5.2
django-js-asset==1.2.2
django-richtextfield==1.6
django-environ==0.4.5
django-heroku==0.3.1
gunicorn==20.0.4
idna==2.10
Jinja2==2.11.2
psycopg2==2.8.6
python-dateutil==2.8.1
python-decouple==3.4
pytz==2020.1
requests==2.24.0
six==1.15.0
urllib3==1.25.9
